+++ CHANGELOG.md
@@ -8,6 +8,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- Handling invalid xPath expression warnings
+  ([#361](https://github.com/jjriv/emogrifier/pull/361))
 
 ### Changed
 
@@ -24,8 +26,6 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 
 ### Security
-
-
 
 ## 1.1.0 (2016-09-18)